+++ csatar-plugins/csatar/models/PermissionBasedAccess.php
@@ -217,6 +217,11 @@ class PermissionBasedAccess extends Model
         return '';
     }
 
+    public static function getEagerLoadSettings(string $useCase = null): array
+    {
+        return [];
+    }
+
     public function getAssociation() {
         return null;
     }