+++ Changelog.md
@@ -4,6 +4,8 @@
 
 - CS-411 - Fixed the header on the team page, so neither the image, nor the header will overflow the screen
 
+- CS-499 - Optimized pages with forms, added eager load settings to organization base models, optimized organization base models to reduce query numbers
+
 ## 1.9.0
 ### 2023-02-17